requirements.txt
Django==3.2.20
django-mysql==4.11.0
mysqlclient==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt  
django-cors-headers
django-environ
gunicorn

